fix: Push Changes UI command forces configs to be built.
ZEN-34837
