+++ Products/ZenCollector/configcache/invalidator.py
@@ -10,7 +10,6 @@
 from __future__ import print_function, absolute_import
 
 import logging
-import time
 
 from multiprocessing import Process
 
@@ -24,9 +23,15 @@ from Products.Zuul.catalog.interfaces import IModelCatalogTool
 
 from .app import Application
 from .app.args import get_subparser
-from .cache import CacheKey, CacheQuery, ConfigStatus
+from .cache import CacheQuery
 from .debug import Debug as DebugCommand
 from .dispatcher import BuildConfigTaskDispatcher
+from .handlers import (
+    NewDeviceHandler,
+    DeviceUpdateHandler,
+    MissingConfigsHandler,
+    RemoveConfigsHandler,
+)
 from .modelchange import InvalidationCause
 from .utils import (
     get_build_timeout,
@@ -179,7 +184,7 @@ def _removeDeleted(log, store, devices):
         if (key.device, key.monitor) not in devices
     )
     if devices_not_found:
-        _RemoveConfigsHandler(log, store)(devices_not_found)
+        RemoveConfigsHandler(log, store)(devices_not_found)
     else:
         log.info("no dangling configurations found")
 
@@ -190,7 +195,7 @@ def _addNewOrChangedDevices(log, store, dispatcher, dmd, devices):
     # Query the catalog for all devices
     new_devices = 0
     changed_devices = 0
-    handle = _NewDeviceHandler(log, store, dispatcher)
+    handle = NewDeviceHandler(log, store, dispatcher)
     for (deviceId, monitorId), uid in devices.iteritems():
         try:
             device = dmd.unrestrictedTraverse(uid)
@@ -227,10 +232,10 @@ class _InvalidationProcessor(object):
     def __init__(self, log, store, dispatcher):
         self.log = log
         self.store = store
-        self._remove = _RemoveConfigsHandler(log, store)
-        self._update = _DeviceUpdateHandler(log, store, dispatcher)
-        self._missing = _MissingConfigsHandler(log, store, dispatcher)
-        self._new = _NewDeviceHandler(log, store, dispatcher)
+        self._remove = RemoveConfigsHandler(log, store)
+        self._update = DeviceUpdateHandler(log, store, dispatcher)
+        self._missing = MissingConfigsHandler(log, store, dispatcher)
+        self._new = NewDeviceHandler(log, store, dispatcher)
 
     def __call__(self, device, oid, reason):
         uid = device.getPrimaryId()
@@ -258,9 +263,7 @@ class _InvalidationProcessor(object):
                 self._new(device.id, monitor, buildlimit, False)
             else:
                 self._update(keys_with_config, minttl)
-                self._missing(
-                    device.id, monitor, keys_with_config, buildlimit
-                )
+                self._missing(device.id, monitor, keys_with_config, buildlimit)
         elif reason is InvalidationCause.Removed:
             self._remove(keys_with_config)
         else:
@@ -271,156 +274,4 @@ class _InvalidationProcessor(object):
                 device,
                 monitor,
                 oid,
-            )
-
-
-class _NewDeviceHandler(object):
-
-    def __init__(self, log, store, dispatcher):
-        self.log = log
-        self.store = store
-        self.dispatcher = dispatcher
-
-    def __call__(self, deviceId, monitor, buildlimit, newDevice=True):
-        keys = tuple(
-            CacheKey(svcname, monitor, deviceId)
-            for svcname in self.dispatcher.service_names
-        )
-        keys_with_pending_status = set(
-            status.key
-            for status in self.store.get_status(*keys)
-            if isinstance(status, ConfigStatus.Pending)
-        )
-        for key in keys_with_pending_status:
-            self.log.debug(
-                "build job already submitted for this config  "
-                "device=%s collector=%s service=%s",
-                key.device,
-                key.monitor,
-                key.service,
-            )
-        keys_without_pending_status = set(keys) - keys_with_pending_status
-        now = time.time()
-        self.store.set_pending(
-            *((key, now) for key in keys_without_pending_status)
-        )
-        for key in keys_without_pending_status:
-            self.dispatcher.dispatch(
-                key.service, key.monitor, key.device, buildlimit, now
-            )
-            self.log.info(
-                "submitted build job for %s  "
-                "device=%s collector=%s service=%s",
-                "new device" if newDevice else "device with new device class",
-                key.device,
-                key.monitor,
-                key.service,
-            )
-
-
-class _DeviceUpdateHandler(object):
-
-    def __init__(self, log, store, dispatcher):
-        self.log = log
-        self.store = store
-        self.dispatcher = dispatcher
-
-    def __call__(self, keys, minttl):
-        current_statuses = tuple(
-            status
-            for status in self.store.get_status(*keys)
-            if isinstance(status, ConfigStatus.Current)
-        )
-
-        now = time.time()
-        retirement = now - minttl
-
-        retired = set(
-            status.key
-            for status in current_statuses
-            if status.updated >= retirement
-        )
-        expired = set(
-            status.key
-            for status in current_statuses
-            if status.key not in retired
-        )
-
-        self.store.set_retired(*((key, now) for key in retired))
-        self.store.set_expired(*((key, now) for key in expired))
-
-        for key in retired:
-            self.log.info(
-                "retired configuration of changed device  "
-                "device=%s collector=%s service=%s",
-                key.device,
-                key.monitor,
-                key.service,
-            )
-        for key in expired:
-            self.log.info(
-                "expired configuration of changed device  "
-                "device=%s collector=%s service=%s",
-                key.device,
-                key.monitor,
-                key.service,
-            )
-
-
-class _MissingConfigsHandler(object):
-
-    def __init__(self, log, store, dispatcher):
-        self.log = log
-        self.store = store
-        self.dispatcher = dispatcher
-
-    def __call__(self, deviceId, monitor, keys, buildlimit):
-        """
-        @param keys: These keys are associated with a config
-        @type keys: Sequence[CacheKey]
-        """
-        # Send a job for for all config services that don't currently have
-        # an associated configuration.  Some ZenPacks, i.e. vSphere, defer
-        # their modeling to a later time, so jobs for configuration services
-        # must be sent to pick up any new configs.
-        hasconfigs = tuple(key.service for key in keys)
-        noconfigkeys = tuple(
-            CacheKey(svcname, monitor, deviceId)
-            for svcname in self.dispatcher.service_names
-            if svcname not in hasconfigs
-        )
-        # Identify all no-config keys that already have a status.
-        skipkeys = tuple(
-            status.key for status in self.store.get_status(*noconfigkeys)
-        )
-        now = time.time()
-        for key in (k for k in noconfigkeys if k not in skipkeys):
-            self.store.set_pending((key, now))
-            self.dispatcher.dispatch(
-                key.service, key.monitor, key.device, buildlimit, now
-            )
-            self.log.debug(
-                "submitted build job for possibly missing config  "
-                "device=%s collector=%s service=%s",
-                key.device,
-                key.monitor,
-                key.service,
-            )
-
-
-class _RemoveConfigsHandler(object):
-
-    def __init__(self, log, store):
-        self.log = log
-        self.store = store
-
-    def __call__(self, keys):
-        self.store.remove(*keys)
-        for key in keys:
-            self.log.info(
-                "removed configuration of deleted device  "
-                "device=%s collector=%s service=%s",
-                key.device,
-                key.monitor,
-                key.service,
             )
